backend/requirements.txt
django==3.1.3
djangorestframework==3.12.2